Compact BVH nodes (reduce the memory requirement by half).

+++ src/renderer/path_tracer.cuh
@@ -15,7 +15,7 @@
 class PathTracer: public TracerBase {
 private:
     int* _obj_idxs;
-    float4* _nodes;
+    uint4* _nodes;
 protected:
     CompactedObjInfo* obj_info;
     int num_objs;
@@ -26,7 +26,7 @@ protected:
 
     cudaTextureObject_t bvh_leaves;
     cudaTextureObject_t nodes;
-    float4* _cached_nodes;
+    uint4* _cached_nodes;
 
     float* output_buffer;                // output buffer for images
     int* emitter_prims;
